+++ src/models/Account.php
@@ -1,0 +1,26 @@
+<?php
+/**
+ * @link      https://dukt.net/facebook/
+ * @copyright Copyright (c) 2021, Dukt
+ * @license   https://github.com/dukt/facebook/blob/master/LICENSE.md
+ */
+
+namespace dukt\facebook\models;
+
+use craft\base\Model;
+
+/**
+ * Tweet model class.
+ *
+ * @author Dukt <support@dukt.net>
+ * @since  3.0
+ */
+class Account extends Model
+{
+    // Properties
+    // =========================================================================
+
+    public $id;
+
+    public $token;
+}